+++ packages/dash-html-components/dash_html_components/Img.py
@@ -25,12 +25,12 @@ See https://reactjs.org/docs/lists-and-keys.html for more info
 - aria-* (string; optional): A wildcard aria attribute
 - alt (string; optional): Alternative text in case an image can't be displayed.
 - crossOrigin (string; optional): How the element handles cross-origin requests
-- height (string | number; optional): Specifies the height of elements listed here. For all other elements, use the CSS height property.						Note: In some instances, such as <div>, this is a legacy attribute, in which case the CSS height property should be used instead.
+- height (string | number; optional): Specifies the height of elements listed here. For all other elements, use the CSS height property.        Note: In some instances, such as <div>, this is a legacy attribute, in which case the CSS height property should be used instead.
 - sizes (string; optional)
 - src (string; optional): The URL of the embeddable content.
 - srcSet (string; optional): One or more responsive image candidates.
 - useMap (string; optional)
-- width (string | number; optional): For the elements listed here, this establishes the element's width.						Note: For all other instances, such as <div>, this is a legacy attribute, in which case the CSS width property should be used instead.
+- width (string | number; optional): For the elements listed here, this establishes the element's width.        Note: For all other instances, such as <div>, this is a legacy attribute, in which case the CSS width property should be used instead.
 - accessKey (string; optional): Defines a keyboard shortcut to activate or add focus to the element.
 - className (string; optional): Often used with CSS to style elements with common properties.
 - contentEditable (string; optional): Indicates whether the element's content is editable.