+++ packages/dash-html-components/src/components/Link.react.js
@@ -92,7 +92,7 @@ Link.propTypes = {
     'hrefLang': PropTypes.string,
 
     /**
-     * Security Feature that allows browsers to verify what they fetch.
+     * Specifies a Subresource Integrity value that allows browsers to verify what they fetch.
      */
     'integrity': PropTypes.string,