+++ packages/dash-html-components/src/components/Textarea.react.js
@@ -111,6 +111,11 @@ Textarea.propTypes = {
     'form': PropTypes.string,
 
     /**
+     * Provides a hint as to the type of data that might be entered by the user while editing the element or its contents. The attribute can be used with form controls (such as the value of textarea elements), or in elements in an editing host (e.g., using contenteditable attribute).
+     */
+    'inputMode': PropTypes.string,
+
+    /**
      * Defines the maximum number of characters allowed in the element.
      */
     'maxLength': PropTypes.oneOfType([